COL-1922, API for whiteboard-element upsert, which emits to other sessions
