+++ src/api/whiteboard-elements.ts
@@ -1,10 +1,13 @@
 import axios from 'axios'
 import utils from '@/api/api-utils'
 
-export function createWhiteboard(title: string, userIds: number[]) {
-  return axios.post(`${utils.apiBaseUrl()}/api/whiteboard/create`, {title, userIds})
-}
-
 export function deleteWhiteboardElement(socketId: string, uuid: string, whiteboardId: number) {
   return axios.delete(`${utils.apiBaseUrl()}/api/whiteboard/${whiteboardId}/element/${uuid}/delete?socketId=${socketId}`)
+}
+
+export function upsertWhiteboardElements(socketId: string, whiteboardElements: any[], whiteboardId: number) {
+  return axios.post(
+    `${utils.apiBaseUrl()}/api/whiteboard_elements/upsert`,
+    {socketId, whiteboardElements, whiteboardId}
+  )
 }
